Feature/restaurant controller auth

+++ webApi/webApi/Services/RestaurantService.cs
@@ -11,6 +11,7 @@ using webApi.DataTransferObjects.ReviewDTO;
 using webApi.DataTransferObjects.SectionDTO;
 using webApi.Exceptions;
 using webApi.Models;
+using webApi.Enums;
 
 namespace webApi.Services
 {
@@ -49,7 +50,7 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public int CreateNewPositionFromMenu(int id, NewPositionFromMenu newPosition)
+        public int CreateNewPositionFromMenu(int id, NewPositionFromMenu newPosition, int userId)
         {
             if (newPosition is null) throw new BadRequestException("Bad request");
 
@@ -59,7 +60,13 @@ namespace webApi.Services
                 .Sections
                 .FirstOrDefault(s => s.Id == id);
 
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
             if (section is null) throw new NotFoundException("Resources not found");
+            if (user is null || user.RestaurantId != section.RestaurantId) throw new UnathorisedException("Unauthorized");
 
             _context.Dishes.Add(dish);
             _context.SaveChanges();
@@ -95,7 +102,7 @@ namespace webApi.Services
             return restaurant.Id;
         }
 
-        public int CreateSection(int id, string sectionName)
+        public int CreateSection(int id, string sectionName, int userId)
         {
             if (sectionName is null || sectionName == string.Empty) throw new BadRequestException("Bad request");
 
@@ -109,17 +116,32 @@ namespace webApi.Services
                 RestaurantId = id
             };
 
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id))
+                throw new UnathorisedException("Unauthorized");
+
             _context.Sections.Add(section);
             _context.SaveChanges();
 
             return section.Id;
         }
 
-        public void DeactivateRestaurant(int id)
+        public void DeactivateRestaurant(int id, int userId)
         {
             var restaurant = _context
                 .Restaurants
                 .FirstOrDefault(item => item.Id == id);
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || user.RestaurantId != id) throw new UnathorisedException("Unauthorized");
 
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
@@ -127,11 +149,18 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void DeleteRestaurant(int id)
+        public void DeleteRestaurant(int id, int userId)
         {
             var restaurant = _context
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unauthorized");
 
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
@@ -139,22 +168,36 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void DeleteSection(int id)
+        public void DeleteSection(int id, int userId)
         {
             var section = _context.Sections.FirstOrDefault(s => s.Id == id);
 
+            var user = _context
+                    .Users
+                    .Where(u => u.Id == userId)
+                    .FirstOrDefault();
+            
             if (section is null) throw new NotFoundException("Resources not found");
+
+            if (user is null ||  user.RestaurantId != section.Id) throw new UnathorisedException("Unauthorized");
 
             _context.Sections.Remove(section);
             _context.SaveChanges();
         }
 
-        public List<ComplaintDTO> GetAllComplaitsForRestaurants(int? id)
+        public List<ComplaintR> GetAllComplaitsForRestaurants(int? id, int userId)
         {
             var restaurant = _context
                 .Restaurants
                 .Include(item => item.Orders)
                 .FirstOrDefault(item => item.Id == id);
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
 
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
@@ -164,71 +207,115 @@ namespace webApi.Services
                 .Join(_context.Orders.Where(item => item.RestaurantId == id), x => x.OrderId, y => y.Id, (x, y) => x)
                 .ToList();
 
-            var complaintDTOs = _mapper.Map<List<ComplaintDTO>>(complaints);
+            List<ComplaintR> complaintDTOs = new List<ComplaintR>();
+            if (user.Role == (int)Role.Restaurer)
+            {
+                complaintDTOs = _mapper.Map<List<ComplaintR>>(restaurant.Reviews);
+            }
+            else
+            {
+                var rests = _mapper.Map<List<ComplaintDTO>>(restaurant.Reviews);
+                foreach (var rest in rests) complaintDTOs.Add(rest);
+            }
             return complaintDTOs;
         }
 
-        public List<OrderDTO> GetAllOrdersForRestaurants(int id)
+        public List<OrderR> GetAllOrdersForRestaurants(int id, int userId)
         {
             var restaurant = _context
                 .Restaurants
                 .Include(item => item.Orders)
                 .FirstOrDefault(item => item.Id == id);
 
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
+
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
-            var orderDTOs = _mapper.Map<List<OrderDTO>>(restaurant.Orders);
+            var orderDTOs = _mapper.Map<List<OrderR>>(restaurant.Orders);
 
             return orderDTOs;
         }
 
-        public List<RestaurantDTO> GetAllRestaurants()
+        public List<RestaurantC> GetAllRestaurants(int userId)
         {
             var restaurants = _context
                 .Restaurants
                 .Include(item => item.Address)
                 .ToList();
+            
+            var user = _context
+            .Users
+            .Where(u => u.Id == userId)
+            .FirstOrDefault();
 
-            var restaurantDTOs = _mapper.Map<List<RestaurantDTO>>(restaurants);
+            List<RestaurantC> restaurantDTOs = new List<RestaurantC>();
+            if (user.Role == (int)Role.Customer)
+            {
+                restaurantDTOs = _mapper.Map<List<RestaurantC>>(restaurants);
+            }
+            else
+            {
+                var rests = _mapper.Map<List<RestaurantDTO>>(restaurants);
+                foreach (var rest in rests) restaurantDTOs.Add(rest);
+            }
             return restaurantDTOs;
         }
 
-        public List<ReviewDTO> GetAllReviewsForRestaurants(int? id)
+        public List<ReviewR> GetAllReviewsForRestaurants(int? id, int userId )
         {
             var restaurant = _context
                .Restaurants
                .Include(item => item.Reviews)
                .FirstOrDefault(item => item.Id == id);
 
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
+
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
-            var reviewDTOs = _mapper.Map<List<ReviewDTO>>(restaurant.Reviews);
-
+            List<ReviewR> reviewDTOs = new List<ReviewR>();
+            if (user.Role == (int)Role.Restaurer)
+            {
+                reviewDTOs = _mapper.Map<List<ReviewR>>(restaurant.Reviews);
+            }
+            else
+            {
+                var rests = _mapper.Map<List<ReviewDTO>>(restaurant.Reviews);
+                foreach (var rest in rests) reviewDTOs.Add(rest);
+            }
             return reviewDTOs;
         }
 
-        public PositionFromMenuDTO GetDishById(int id)
-        {
-            var dish = _context
-                          .Dishes
-                          .FirstOrDefault(r => r.Id == id);
-
-            if (dish is null) throw new NotFoundException("Resource not found");
-
-            var dishDTO = _mapper.Map<PositionFromMenuDTO>(dish);
-            return dishDTO;
-        }
-
-        public RestaurantDTO GetRestaurantById(int? id)
+        public RestaurantC GetRestaurantById(int? id, int userId)
         {
             var restaurant = _context
                             .Restaurants
                             .Include(item => item.Address)
                             .FirstOrDefault(r => r.Id == id);
 
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
+
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
-            var restaurantDTO = _mapper.Map<RestaurantDTO>(restaurant);
+            RestaurantC restaurantDTO;
+            if (user.Role == (int)Role.Customer)
+                restaurantDTO = _mapper.Map<RestaurantC>(restaurant);
+            else
+                restaurantDTO = _mapper.Map<RestaurantDTO>(restaurant);
             return restaurantDTO;
         }
 
@@ -250,11 +337,18 @@ namespace webApi.Services
 
         }
 
-        public void ReactivateRestaurant(int id)
+        public void ReactivateRestaurant(int id, int userId)
         {
             var restaurant = _context
                .Restaurants
                .FirstOrDefault(item => item.Id == id);
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
 
             if (restaurant is null) throw new NotFoundException("Resource not found");
 
@@ -262,12 +356,30 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void RemovePositionFromMenu(int id)
+        public void RemovePositionFromMenu(int id, int userId)
         {
             var dish = _context.Dishes
                 .FirstOrDefault(d => d.Id == id);
-
+           
             if (dish is null) throw new NotFoundException("Resource not found");
+
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null) throw new UnathorisedException("Unathourized");
+
+            var section = _context
+                .Sections
+                .Where(s => s.Id == dish.SectionId)
+                .FirstOrDefault();
+
+            if (section is null) throw new NotFoundException("Resource not found");
+
+            if ((user.Role == (int)Role.Restaurer && user.RestaurantId != section.RestaurantId)) throw new UnathorisedException("Unauthorized");
+
 
             var orderDishes = _context.OrderDishes.Where(item => item.DishId == id);
             _context.OrderDishes.RemoveRange(orderDishes);
@@ -276,9 +388,24 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void SetFavouriteRestaurant(int id)
+        public void SetFavouriteRestaurant(int id, int userId)
         {
-            throw new NotImplementedException();
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            var restaurant = _context
+                .Restaurants
+                .Where(r => r.Id == id)
+                .FirstOrDefault();
+
+            if (restaurant is null) throw new NotFoundException("Resource not found");
+
+            if (user is null || user.Role != (int)Role.Customer) throw new UnathorisedException("Unathourized");
+
+            user.RestaurantId = id;
+            _context.SaveChanges();
         }
 
         public void UnblockRestaurant(int id)
@@ -293,7 +420,7 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void UpdatePositionFromMenu(int id, NewPositionFromMenu newPosition)
+        public void UpdatePositionFromMenu(int id, NewPositionFromMenu newPosition, int userId)
         {
             if (newPosition is null) throw new BadRequestException("Bad request");
 
@@ -301,6 +428,13 @@ namespace webApi.Services
                 .FirstOrDefault(d => d.Id == id);
 
             if (dish is null) throw new NotFoundException("Resource not found");
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
 
             dish.Name = newPosition.Name;
             dish.Price = newPosition.Price;
@@ -310,13 +444,20 @@ namespace webApi.Services
             _context.SaveChanges();
         }
 
-        public void UpdateSection(int id, string newSectionName)
+        public void UpdateSection(int id, string newSectionName, int userId)
         {
             if (newSectionName is null || newSectionName == string.Empty) throw new BadRequestException("Bad request");
 
             var section = _context.Sections.FirstOrDefault(s => s.Id == id);
 
             if (section is null) throw new NotFoundException("Resources not found");
+
+            var user = _context
+                .Users
+                .Where(u => u.Id == userId)
+                .FirstOrDefault();
+
+            if (user is null || (user.Role == (int)Role.Restaurer && user.RestaurantId != id)) throw new UnathorisedException("Unathourized");
 
             section.Name = newSectionName;
             _context.SaveChanges();
